requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0

# Optional for development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0